requirement.txt
absl-py==2.2.2
altair==5.5.0
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
flatbuffers==25.2.10
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
importlib-metadata==7.2.1
jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes
namex==0.0.8
narwhals==1.35.0
numpy
opt-einsum==3.4.0
optree==0.15.0
packaging==23.2
pandas==2.2.3
pillow==10.4.0
protobuf
pyarrow==19.0.1
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.24.0
setuptools==78.1.0
six==1.17.0
smmap==5.0.2
streamlit
tenacity==8.5.0
tensorboard
tensorboard-data-server==0.7.2
tensorflow
tensorflow-intel
termcolor==3.0.1
toml==0.10.2
tornado==6.4.2
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
validators==0.34.0
watchdog==6.0.0
werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
zipp==3.21.0
Flask

tf_keras
